requirements.txt
flask==2.3.3
flask-jwt-extended==4.5.2
# python-ldap==3.4.3  # Requires system dependencies - install manually if needed
python-dotenv==1.0.0
pyyaml==6.0.1
pytest==7.4.0
pytest-cov==4.1.0
requests==2.31.0
